backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
sqlparse==0.5.3
psycopg2-binary==2.9.9
python-dotenv
djangorestframework-simplejwt
drf-spectacular
django-cors-headers==4.4.0
pyproj
django-simple-history
djangorestframework-api-key